+++ engine/list_collection/list.hpp
@@ -4,22 +4,8 @@
 
 #pragma once
 
-#include <algorithm>
-#include <cassert>
-#include <cstdint>
-#include <thread>
-#include <unordered_map>
-#include <unordered_set>
-
-#include "../alias.hpp"
-#include "../collection.hpp"
 #include "../dl_list.hpp"
-#include "../hash_table.hpp"
-#include "../lock_table.hpp"
-#include "../structures.hpp"
-#include "../utils/utils.hpp"
-#include "../write_batch_impl.hpp"
-#include "kvdk/engine.hpp"
+#include "kvdk/types.hpp"
 
 namespace KVDK_NAMESPACE {
 class ListIteratorImpl;
@@ -157,92 +143,5 @@ class List : public Collection {
   DLList dl_list_;
   PMEMAllocator* pmem_allocator_;
   std::atomic<size_t> size_;
-};
-
-class ListIteratorImpl final : public ListIterator {
- public:
-  ListIteratorImpl(Engine* engine, List* list, const SnapshotImpl* snapshot,
-                   bool own_snapshot)
-      : engine_(engine),
-        list_(list),
-        snapshot_(snapshot),
-        own_snapshot_(own_snapshot),
-        dl_iter_(&list->dl_list_, list->pmem_allocator_, snapshot) {}
-
-  void Seek(long index) final {
-    if (index < 0) {
-      SeekToLast();
-      long cur = -1;
-      while (cur-- > index && Valid()) {
-        Prev();
-      }
-    } else {
-      SeekToFirst();
-      long cur = 0;
-      while (cur++ < index && Valid()) {
-        Next();
-      }
-    }
-  }
-
-  void SeekToFirst() final { dl_iter_.SeekToFirst(); }
-
-  void SeekToLast() final { dl_iter_.SeekToLast(); }
-
-  void SeekToFirst(StringView elem) final {
-    SeekToFirst();
-    Next(elem);
-  }
-
-  void SeekToLast(StringView elem) final {
-    SeekToLast();
-    Prev(elem);
-  }
-
-  bool Valid() const final { return dl_iter_.Valid(); }
-
-  void Next() final { dl_iter_.Next(); }
-
-  void Prev() final { dl_iter_.Prev(); }
-
-  void Next(StringView elem) final {
-    while (Valid()) {
-      Next();
-      if (!Valid() || equal_string_view(elem, dl_iter_.Value())) {
-        break;
-      }
-    }
-  }
-
-  void Prev(StringView elem) final {
-    while (Valid()) {
-      Prev();
-      if (!Valid() || equal_string_view(elem, dl_iter_.Value())) {
-        break;
-      }
-    }
-  }
-
-  std::string Elem() const final {
-    if (!Valid()) {
-      kvdk_assert(false, "Accessing data with invalid ListIterator!");
-      return std::string{};
-    }
-    auto sw = dl_iter_.Value();
-    return std::string{sw.data(), sw.size()};
-  }
-
-  ~ListIteratorImpl() final {
-    if (own_snapshot_ && snapshot_) {
-      engine_->ReleaseSnapshot(snapshot_);
-    }
-  }
-
- private:
-  Engine* engine_;
-  List* list_;
-  const SnapshotImpl* snapshot_;
-  bool own_snapshot_;
-  DLListDataIterator dl_iter_;
 };
 }  // namespace KVDK_NAMESPACE
